[KYUUBI #1386] Refactor SparkThriftBinaryFrontendService to extract connectionUrl into super class
